==> 2_ScrapingMoviePages.ipynb <==
# Scrape Raw Data from Movie Review Pages
#### In this notebook, a list of URL's was imported and used to scrape individual movie pages for text and non-text data. Data then run through Beautiful Soup to strip away html artifacts and exported for further processing.
#### Separate functions extract each individual movie feature to facilitate trouble-shooting and provide for maximum modularity.

### Importing URL's

import pandas as pd
import seaborn as sns
import requests, re, json, time, copy
from bs4 import BeautifulSoup as bs4

### Load list of movie URL's and visit them to scrape movie features

# Load json of movies_urls
with open('data/movies_urls.json') as json_file:  
    movies_urls = json.load(json_file)

movies_urls[:5]

### Main functions used to scrape data

movies_features = []
def scrape_movie_features(first_movie, num_movies_to_scrape):
    movie_features = {}
    missed_movies = {}
    for movie in range(num_movies_to_scrape):
        movie = (movie + first_movie)
        url = movies_urls[movie]
        res = requests.get(url)
        soup = bs4(res.content, 'lxml')
        print(movie, url[47:])
        if res.status_code == 200:
            movie_features = get_movie_features(movie, url, soup)
            movies_features.append(movie_features)
        else:
            missed_movies.append([movie, url])
        if movie == len(movies_urls):
            print('That was the last movie.')
            break
    return movies_features, missed_movies

movies_features = []
def get_movie_features(num, url, soup):
    movie_features = {}
    movie_features['movie_id'] = num
    movie_features['slug'] = url[47:]
    movie_features['title'] = get_title(soup)
    movie_features['age'] = get_age(soup)
    movie_features['family_topics'] = get_family_topics(soup)
    movie_features['is_it_any_good'] = get_is_it_any_good(soup)
    movie_features['movie_details_raw'] = get_movie_details_raw(soup)
    movie_features['one_line_description'] = get_one_line_description(soup)
    movie_features['overall_rating'] = get_overall_rating(soup)
    movie_features['parental_rating_and_spoilers_raw'] = get_parental_rating_and_spoilers_raw(soup)
    movie_features['what_is_the_story_raw'] = get_what_is_the_story(soup)
    movie_features['what_parents_need_to_know_raw'] = get_what_parents_need_to_know(soup)
    return movie_features

### Helper Functions to get specific raw movie features

def get_title(soup):
    return str(soup.h1)[4:-5]

def get_age(soup):
    return str(soup.find('div', 'csm-green-age').get_text())

def get_family_topics(soup):
    return str(soup.find('div',
                         'field-name-field-family-topics').get_text())   

def get_is_it_any_good(soup):
    return str(soup.find('div', 'field-name-field-any-good').get_text())

def get_movie_details_raw(soup):
    return str(soup.find('div', 'pane-product-details').get_text())   

def get_one_line_description(soup):
    return str(soup.find('div', 'field-name-field-one-liner').get_text())

def get_overall_rating(soup):
    rating = re.search("\d", str(soup.find('div', 'field_stars_rating')))
    return str(rating[0])

def get_parental_rating_and_spoilers_raw(soup):
    return str(soup.find_all('div', 'field-type-field-collection'))

def get_what_is_the_story(soup):
    return str(soup.find('div',
                         'field-name-field-what-is-story').get_text())

def get_what_parents_need_to_know(soup):
    return str(soup.find('div', 'field-name-field-parents-need-to-know'))

### Scraping Movie Features

# DO NOT RUN THIS CELL  will run a scraper that will take almost three hours
movies_features, missed_movies = scrape_movie_features(0, 10000)

len(movies_features)

# Run if you'd like to see the scraped raw, unprocessed movie data
movies_features[len(movies_features) - 1]

# replace RENUMBER with len(movies_features)-1 and edit future notebooks to update system
# to be able to recommend more recent movies. See below for further explanation
movies_features_0_to_RENUMBERnew = movies_features

with open('data/movies_features_0_to_RENUMBERnew.json', 'w') as output:
    json.dump(movies_features_0_to_RENUMBERnew, output)

#### Movie features saved to json file for further cleaning and data exploration in Notebook 3.
