==> 3_Extracting_Features.ipynb <==
## Eliminate Duplicates and Extract Movie Features from Scraped Data
#### In this notebook, we take movies from the previous notebook, identify and eliminate duplicate movies, and further clean and extract important movie features and rearrange them to facilitate analysis. Extracted features are then saved for further cleaning in Notebook 4.

import pandas as pd
import seaborn as sns
import requests, re, json, time
from bs4 import BeautifulSoup as bs4

### Import movies_features files

# Load json of movies_features. N.B., this notebook is set up to use data scraped previously
# and unfortunately too large to push to GitHub. If you'd like to replicate this process,
# go to https://drive.google.com/open?id=19VPd3pdjrwVV56OVDNtKo2o_V3afwrvU to get this file
# and put in the /data file.
with open('data/movies_features_0_to_8764new.json') as json_file:  
    movies_features = json.load(json_file)

# Run cell if yoU want to see movies_features of first 3 movies
movies_features[0:2]

### Identify and Eliminate Duplicate Movies
#### Data placed into a pandas DataFrame to facilitate efficient examination of features and identification of duplicate movies. Of the 8778 movies in our collection, 8625 are unique (98.3%). 153 movies dropped.

df = pd.DataFrame(movies_features)

df.head()

len(df)

len(df.title.unique())

# 99.4% of movies are unique.
8714/8765

df.drop_duplicates('title', inplace=True)

len(df) 

movies_features = df.to_dict('records')

# Run cell if you want to see first 3 movies from movies_features
movies_features[0:2]

## Partial Cleaning and Extracting of Movie Features:
#### Some of the features scraped in the previous notebook need to be extracted in order to flatten movie ratings and other details that may be of importance to family members (e.g., cast, director, movie run time, etc.). I will also prepare text to be used to create a corpus of words used to describe the movie. This will be useful in determining what movies are similar to one another for our recommender.

### Helper Functions to Extract Movie Details, Parental Ratings, etc.:

def extract_movie_details(movie_details):
    details = []
    values = []
    details_lines = re.split('\n', movie_details)
    for line in details_lines:
        try:
            detail_name, detail_value, temp = re.split(':(.*)', line)
            details.append(detail_name)
            values.append(detail_value)
        except:   # if there is not a : in the line, it will ignore line
            pass
    
    return dict(zip(details, values))

def extract_parental_ratings_and_spoilers(parental_ratings):
    soup = bs4(parental_ratings, 'lxml')
    ratings = [re.findall('\d', str(x))[0] for x in soup.find_all('div', attrs={'class':'content-grid-rating'})]
    labels = [x.get_text() for x in soup.find_all('span', attrs={'class':"sprite-cover"})]
    spoilers = [x.get_text() for x in soup.find_all('div', attrs={'class': 'field-type-text-long'})]
    return dict(zip(labels, ratings)), spoilers   

def text_what_parents_need_to_know(what_parents_need_to_know):
    soup = bs4(what_parents_need_to_know, 'lxml')
    what_parents_need_to_know_text = soup.get_text()
    
    return what_parents_need_to_know_text

def extract_features_from_a_movie(movie_num):
    extracted_feat = {}
    movie_id = movies_features[movie_num]['movie_id']
    movie_details_dict = extract_movie_details(movies_features[movie_num]['movie_details_raw'])
    parent_ratings_dict, spoilers = extract_parental_ratings_and_spoilers(movies_features[movie_num]['parental_rating_and_spoilers_raw'])
    what_parents_need_to_know_text = text_what_parents_need_to_know(movies_features[movie_num]['what_parents_need_to_know_raw'])
    
    extracted_feat['movie_id'] = movie_id
    extracted_feat['movie_details_dict'] = movie_details_dict
    extracted_feat['parent_ratings_dict'] = parent_ratings_dict
    extracted_feat['spoilers'] = spoilers
    extracted_feat['what_parents_need_to_know_text'] = what_parents_need_to_know_text
    extracted_feat['family_topics'] = movies_features[movie_num]['family_topics']
    extracted_feat['is_it_any_good'] = movies_features[movie_num]['is_it_any_good']
    extracted_feat['what_is_the_story'] = movies_features[movie_num]['what_is_the_story_raw']
    extracted_feat['age'] = movies_features[movie_num]['age']
    extracted_feat['one_line_description'] = movies_features[movie_num]['one_line_description']
    extracted_feat['overall_rating'] = movies_features[movie_num]['overall_rating']
    extracted_feat['slug'] = movies_features[movie_num]['slug']
    extracted_feat['title'] = movies_features[movie_num]['title']
    return extracted_feat

# Run cell if you want to see example of features extracted from a single movie.
extract_features_from_a_movie(0)

### Extract Features from all movies

extracted_movies = []
def extract_features_from_many_movies(first_movie, num_movies_to_extract):
    '''
    Follow DocString Conventions and annotate functions
    '''
    movie_features = {}
    for movie in range(num_movies_to_extract):
        movie = (movie + first_movie)
        if movie % 100 == 0:
            print(movie)
        extracted_features = extract_features_from_a_movie(movie)
        extracted_movies.append(extracted_features)
    return extracted_movies

# Comment out unless you want to extract features from movies again.
# If you have run previous cells and now have a different number of movies to extract,
# adjust the num_movies_to_extract appropriately.
extracted_movies = extract_features_from_many_movies(0, len(df))

# Comment out cell to supress extra text. Run if you'd like to see extracted features
extracted_movies[-1]

### Export Extracted Movie Features for Further Cleaning and NLP

# # Uncomment if you have run prior notebooks and want to update movie recommender system
# with open('data/extracted_movies.json', 'w') as output:
#     json.dump(extracted_movies, output) 

#### Extracted features from movies saved to json file for final cleaning and division into text and non-text features in the next notebook.
